[FIX] mail: enable composer after message sent
When a message is sent, the composer sent button is disabled (to prevent
multiple emissions) but then was not enabled back.

It is not much of an issue in the basic composer since ENTER key still
sent message, but in extended composer CTRL+ENTER has to be used (and
is less easily known).

issue found when checking opw-1904029
closes #28475